+++ packages/pancake-sass/src/sass.js
@@ -29,32 +29,102 @@ import { Log, Style, WriteFile } from '@gov.au/pancake';
 // Default export
 //--------------------------------------------------------------------------------------------------------------------------------------------------------------
 /**
- * Generate Sass code for a module and it’s dependencies
+ * Get the include path for a sass partial
  *
- * @param  {string} location     - The location of the module to be compiled
- * @param  {object} dependencies - The dependencies of this module
+ * @param  {string} module       - The module name
+ * @param  {object} modules      - An object of all modules and their settings
+ * @param  {string} baseLocation - The current base path
+ * @param  {string} npmOrg       - The npm org scope
  *
- * @return {string}              - Sass code to tie dependencies and module together
+ * @return {string}              - The path to the sass partial
  */
-export const GenerateSass = ( location, dependencies ) => {
-	let sass = ``; //the code goes here
+const GetPath = ( module, modules, baseLocation, npmOrg ) => {
+	let modulePath = '';
 
-	if( dependencies ) {
-		if( Object.keys( dependencies ).length ) {
-			const baseLocation = Path.normalize(`${ location }/../`);
+	for( const item of modules ) {
+		if( item.name === module ) {
+			const moduleName = module.replace(`${ npmOrg }/`, '');
 
-			for( const dependency of Object.keys( dependencies ) ) {
-				const modulePath = dependency.split('/')[ 1 ];
+			modulePath = Path.normalize(`${ baseLocation }/${ moduleName }/${ item.pancake['pancake-module'].sass.path }`);
 
-				sass += `@import "${ Path.normalize(`${ baseLocation }/${ modulePath }/lib/sass/_module.scss`) }";\n`;
-				//TODO: change this to use the modulePackage.pancake['pancake-module'].sass.path value
+			break;
+		}
+	}
+
+	return modulePath;
+}
+
+
+/**
+ * Look up all dependencies of a module by calling yourself
+ *
+ * @param  {string}  module    - The name of the module
+ * @param  {object}  modules   - All modules in an object array
+ * @param  {string}  parent    - The name of the parent module, Defaults to the module argument
+ * @param  {integer} iteration - The depth of the iteration, defaults to 1
+ *
+ * @return {object}            - An object array of the dependencies that are needed for the module
+ */
+const GetDependencies = ( module, modules, parent = module, iteration = 1 ) => {
+	Log.verbose(`Sass: Looking up dependencies at level ${ Style.yellow( iteration ) }`);
+
+	let allDependencies = {};
+
+	if( iteration > 50 ) {
+		Log.error(`Sass: Looks like we found a circular dependency that seems to go no-where from ${ Style.yellow( parent ) }.`);
+	}
+	else {
+
+		for( const item of modules ) {
+			if( item.name === module ) {
+				if( item.peerDependencies ) {
+					for( const subDependency of Object.keys( item.peerDependencies ) ) {
+						const subDependencies = GetDependencies( subDependency, modules, parent, ( iteration + 1 ) );
+
+						allDependencies = Object.assign( allDependencies, subDependencies );
+					}
+				}
+
+				allDependencies = Object.assign( allDependencies, item.peerDependencies );
+
+				break;
 			}
 		}
 
 	}
 
-	sass += `@import "${ Path.normalize(`${ location }/lib/sass/_module.scss`) }";\n`;
-	//TODO: change this to use the modulePackage.pancake['pancake-module'].sass.path value
+	return allDependencies;
+};
+
+
+/**
+ * Generate Sass code for a module and it’s dependencies
+ *
+ * @param  {string} location - The location of the module to be compiled
+ * @param  {object} name     - The name of the module
+ * @param  {object} modules  - All modules and their dependencies
+ * @param  {object} npmOrg   - The name of the npm org scope
+ *
+ * @return {string}          - Sass code to tie dependencies and module together
+ */
+export const GenerateSass = ( location, name, modules, npmOrg ) => {
+	let sass = ``; //the code goes here
+
+	const baseLocation = Path.normalize(`${ location }/../`);
+	const dependencies = GetDependencies( name, modules );
+
+	Log.verbose(`Sass: For ${ Style.yellow( name ) } we found the following dependencies ${ Style.yellow( JSON.stringify( dependencies ) ) }`);
+
+	if( dependencies ) {
+		for( const dependency of Object.keys( dependencies ) ) {
+			const modulePath = GetPath( dependency, modules, baseLocation, npmOrg )
+
+			sass += `@import "${ modulePath }";\n`;
+		}
+	}
+
+	const modulePath = GetPath( name, modules, baseLocation, npmOrg )
+	sass += `@import "${ modulePath }";\n`;
 
 	return sass;
 };