+++ packages/pancake-js/CHANGELOG.md
@@ -6,10 +6,30 @@ Pancake JS PLUGIN
 
 ## Versions
 
+* [v1.0.6 - Receiving global settings from pancake now](v106)
+* [v1.0.5 - Loading fixes](v105)
+* [v1.0.4 - Bug hunting](v104)
 * [v1.0.3 - 💥 Initial version](v103)
 
 
 ----------------------------------------------------------------------------------------------------------------------------------------------------------------
+
+
+## v1.0.6
+
+- Receiving global settings from pancake now
+- Loading now separate from main module in case the main module is globally installed
+
+
+## v1.0.5
+
+- Loading as a plugin is now fixed, even when you have pancake installed globally
+- Removed pancake as a dependency
+
+
+## v1.0.4
+
+- Fixed some bugs
 
 
 ## v1.0.3